날씨 삭제 스케쥴 데이터 1개라도 있을 경우 삭제 로직 추가

+++ src/main/java/com/jagiya/weather/service/WeatherService.java
@@ -499,8 +499,10 @@ public class WeatherService {
         String baseDate = LocalDate.now().minusDays(1).format(DateTimeFormatter.ofPattern("yyyyMMdd"));
         List<Weather> weatherList = weatherRepository.findByBaseDateLessThanEqual(baseDate);
         if (weatherList != null) {
-            log.info("deleteBeforeWeather Size {}, baseTime {}", weatherList.size() , baseDate);
+            if (weatherList.size() > 0) {
+                log.info("deleteBeforeWeather Size {}, baseTime {}", weatherList.size() , baseDate);
+                weatherRepository.deleteAll(weatherList);
+            }
         }
-        weatherRepository.deleteAll(weatherList);
     }
 }
